requirements.txt
pandas===2.3.1
sqlalchemy===2.0.42
psycopg2-binary==2.9.10
faker===37.5.3
requests==2.32.4
plotly==6.2.0
streamlit==1.48.0
